+++ src/objects.js
@@ -84,7 +84,7 @@ BlockEditorMorph, BlockDialogMorph, PrototypeHatBlockMorph,  BooleanSlotMorph,
 localize, TableMorph, TableFrameMorph, normalizeCanvas, VectorPaintEditorMorph,
 AlignmentMorph, Process, WorldMap, copyCanvas, useBlurredShadows*/
 
-modules.objects = '2020-December-03';
+modules.objects = '2020-December-04';
 
 var SpriteMorph;
 var StageMorph;
@@ -978,13 +978,6 @@ SpriteMorph.prototype.initBlocks = function () {
             category: 'sensing',
             spec: '%rel to %dst',
             defaults: [['distance'], ['mouse-pointer']]
-        },
-        reportDistanceFacing: { // experimental, under construction
-            only: SpriteMorph,
-            type: 'reporter',
-            category: 'sensing',
-            spec: 'distance facing %cln',
-            defaults: [['myself']]
         },
         doResetTimer: {
             type: 'command',
@@ -2547,8 +2540,6 @@ SpriteMorph.prototype.blockTemplates = function (category) {
             blocks.push(block('reportThreadCount'));
             blocks.push(block('reportStackSize'));
             blocks.push(block('reportFrameCount'));
-            blocks.push('-');
-            blocks.push(block('reportDistanceFacing'));
         }
 
 	/////////////////////////////////